Initialize axios with base Url
