+++ src/API/asyncFetchCategories.js
@@ -1,0 +1,20 @@
+import axios from "axios";
+import { fetchCategoriesFailure, fetchCategoriesSuccess } from "../actions";
+import categories from "../reducers/categories";
+import { listAllMealCategories } from './apiUrls';
+
+const fetchCategories = () => {
+  return function(dispatch) {
+    dispatch(fetchCategoriesRequest);
+    axios.get(listAllMealCategories)
+    .then(response => {
+      const categories = response.categories.map(category => category);
+      dispatch(fetchCategoriesSuccess(categories));
+    })
+    .catch(error => {
+      dispatch(fetchCategoriesFailure(error.message));
+    })
+  }
+}
+
+export default fetchCategories
