+++ src/API/baseHttp.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+export default axios.create({
+  baseURL: "https://www.themealdb.com/api/json/v1/1",
+  headers: {
+    "Content-type": "application/json"
+  }
+});
